LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_element_haspointercapture_release_pending_capture-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Element.hasPointerCapture test after the pending pointer capture element releases pointer capture

Test Description: This test checks if Element.hasPointerCapture returns value correctly after the pending pointer capture element releases pointer capture
Press black rectangle and do not release
Move your pointer to purple rectangle
Release the pointer

The following pointer types were detected: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN hasPointerCapture test after the pending pointer capture element releases pointer capture
